updated Learner interface added enums for gridtypes and SLE solvers fixed conversions warnings
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2078 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/algorithm/DMSystemMatrixSPVectorizedIdentity.cpp
@@ -109,7 +109,7 @@ void DMSystemMatrixSPVectorizedIdentity::mult(sg::base::DataVectorSP& alpha, sg:
     this->computeTimeMultTrans += this->B->multTransposeVectorized(temp, result);
     this->completeTimeMultTrans += this->myTimer->stop();
 
-    result.axpy(numTrainingInstances*this->lamb, alpha);
+    result.axpy(((float)numTrainingInstances)*this->lamb, alpha);
 }
 
 void DMSystemMatrixSPVectorizedIdentity::generateb(sg::base::DataVectorSP& classes, sg::base::DataVectorSP& b)
